requirements.txt
langchain
langchain-community
langchain-openai
langchain-postgres
dateparser
langgraph
PyPDF2
psycopg2-binary
pgvector
pdfplumber
pytesseract
pillow
python-dotenv
sqlalchemy
PyMuPDF
azure-ai-documentintelligence
azure-identity
langsmith